chore: Fix example App.tsx for TypeScript (#135)
Hi @Rapsssito!

It is an awesome work of all colaborators! Yesterday I was able to test and this example seems to be a little deprecated, in that line:
`{global.HermesInternal == null ? null : (`
instead of:
` {global.HermesInternal == null ? null : (`

So, based on this issue https://github.com/react-native-community/react-native-template-typescript/pull/78/files

I propose to modify to:
`{!usingHermes ? null : (`

And including the usingHermes declaration line also: 
`  const usingHermes = typeof HermesInternal === 'object' && HermesInternal !== null;`


Moreover, I have proposed to a Functional Component, which it seems it is more performative and inserted the .tsx to match to the more recent typing.

I hope this should improve the startup of the running example! Thanks for the oportunity to contribute a little as you have done a bigger!


Best Regards!